+++ src/main/java/com/CuteNekoDragon/Core/common/data/SVOTags.java
@@ -1,0 +1,25 @@
+package com.CuteNekoDragon.Core.common.data;
+
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
+import net.minecraftforge.registries.ForgeRegistries;
+
+import com.CuteNekoDragon.Core.SVOCore;
+
+public final class SVOTags {
+
+    public static final class Items {
+
+        public static final TagKey<Item> Minerals = createItemTag("minerals");
+        public static final TagKey<Item> Gemstones = createItemTag("gemstones");
+
+        private static TagKey<Item> createItemTag(String path) {
+            return createItemTag(SVOCore.id(path));
+        }
+
+        private static TagKey<Item> createItemTag(ResourceLocation resLoc) {
+            return TagKey.create(ForgeRegistries.ITEMS.getRegistryKey(), resLoc);
+        }
+    }
+}